IconHelper leaves unsed API allocated memory
Your working with the Windows API is dirty. You don't frees API allocated memory. In description of ExtractIconEx states under "Remarks" you have to destroy the icons, that means calling DestroyIcon for each icon handle (your IntPtr's).

I'm not familiar right now with GIT so I don't fix it in code, but introduces a "TODO" comment with links to the MSDN descriptions of both functions.