Add root as default parent for new backups.
This gets overwritten for anything not in the drive root, so newly-copied backups in the drive root have their parent explicitly set to root, instead of inheriting it from the source file.
Apparently this is how the Google Drive API treats copies with File unset or set to an empty list: the newly-created copy inherits the Parents list of the source file it was copied from.

This will fix a future incompatibility between root backups and shortcuts.
If a shortcut exists in the root, and we try to back it up, the backups should appear alongside the shortcut.
However, unless we set the parent to root, they would inherit the target's parent directory and appear alongside the target in it's home directory (wherever in the drive that might be).